Updating Application's POM file to import Kafka, Solace and Local Storage Plugins.
This ensures that these plugins are available to the main application so it can still perform scans. Without these here then the plugins would need to get built separately and deployed with the EMA.

These dependencies will get removed from the Application POM file in the future when we can do additional JAR deployments with it.